@@ -1,12 +1,14 @@
 """
 Metrics normalizer stub
 """
-from typing import List, Dict, Any
+
+from typing import Any
+
 
 class MetricsNormalizer:
     """Simple metrics normalizer for testing"""
-    
-    def normalize(self, data: List[Dict[str, Any]]) -> List[Dict[str, Any]]:
+
+    def normalize(self, data: list[dict[str, Any]]) -> list[dict[str, Any]]:
         """Normalize metrics data"""
         # Simple pass-through for now
         return data
