@@ -5,7 +5,7 @@ This package provides functionality to execute predefined cases from CSV files
 and collect standardized metrics, not to design experiments.
 
 Core Components:
-- Runner: Main execution engine 
+- Runner: Main execution engine
 - Executor: Case execution management
 - CSVHandler: Deterministic CSV I/O
 - CaseCache: Result caching with case_hash
@@ -15,22 +15,23 @@ __version__ = "0.1.0"
 __author__ = "Strataregula Team"
 __email__ = "support@strataregula.dev"
 
-from .core.runner import Runner, ExecutionResult
-from .core.executor import CaseExecutor
-from .io.csv_handler import CSVHandler
 from .adapters.base import BaseAdapter
+from .core.executor import CaseExecutor
+from .core.runner import ExecutionResult, Runner
+from .io.csv_handler import CSVHandler
 
 # Plugin integration
 try:
     from .plugin import DOERunnerPlugin
+
     _PLUGIN_AVAILABLE = True
 except ImportError:
-    DOERunnerPlugin = None
     _PLUGIN_AVAILABLE = False
+    DOERunnerPlugin = None  # type: ignore
 
 __all__ = [
     "Runner",
-    "ExecutionResult", 
+    "ExecutionResult",
     "CaseExecutor",
     "CSVHandler",
     "BaseAdapter",
